refactor into wallet + use fees api
